+++ examples/data-sources/hopsworksai_cluster/data-source.tf
@@ -1,0 +1,4 @@
+
+data "hopsworksai_clusters" "cluster" {
+  cluster_id = "CLUSTER ID"
+}