+++ hopsworksai/provider.go
@@ -41,6 +41,7 @@ func Provider(version string) func() *schema.Provider {
 				},
 			},
 			DataSourcesMap: map[string]*schema.Resource{
+				"hopsworksai_cluster":                                  dataSourceCluster(),
 				"hopsworksai_clusters":                                 dataSourceClusters(),
 				"hopsworksai_aws_instance_profile_policy":              dataSourceAWSInstanceProfilePolicy(),
 				"hopsworksai_azure_user_assigned_identity_permissions": dataSourceAzureUserAssignedIdentityPermissions(),