+++ CHANGELOG.md
@@ -7,3 +7,4 @@ FEATURES:
 * **New Data Source**: `hopsworksai_clusters`
 * **New Data Source**: `hopsworksai_aws_instance_profile_policy`
 * **New Data Source**: `hopsworksai_azure_user_assigned_identity_permissions`
+* **New Data Source**: `hopsworksai_cluster`